Breaking change: Rename convert.py to _convert.py. Import explicitly in __init__.py rather than using `*`.
