+++ excel_dates/__init__.py
@@ -1,1 +1,14 @@
-from .convert import *
+# Let all or most files start with _ to designate them as internal, and only import here the things which are
+#   used by other packages. This helps us to easily know what constitutes a breaking change, and what does not.
+
+from ._convert import (
+    ensure_excel_date,
+    ensure_python_date,
+    ensure_excel_datetime,
+    ensure_python_datetime,
+    ensure_python_time,
+    get_compensation_from_python_to_excel,
+    get_compensation_from_excel_to_python,
+    AnyDateType,
+    epoch,
+)
